refact: rename Signup to Member for consistency
