+++ src/test/kotlin/com/dobby/backend/application/usecase/auth/GenerateTokenWithRefreshTokenUseCaseTest.kt
@@ -1,6 +1,5 @@
-package com.dobby.backend.application.usecase
+package com.dobby.backend.application.usecase.auth
 
-import com.dobby.backend.application.usecase.auth.GenerateTokenWithRefreshTokenUseCase
 import com.dobby.backend.domain.gateway.MemberGateway
 import com.dobby.backend.domain.gateway.TokenGateway
 import com.dobby.backend.domain.model.member.Member
