+++ src/test/kotlin/com/dobby/backend/application/usecase/member/CreateResearcherUseCaseTest.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.application.usecase.signup
+package com.dobby.backend.application.usecase.member
 
 import io.kotest.core.spec.style.BehaviorSpec
